fix: fixed hexponent class filter works as intended

+++ Application_1/library/include/EFPDF.h
@@ -8,7 +8,7 @@ private:
     std::string histogramPath;
     std::string arguments;
     double gmin;
-    int alpha;
+    double alpha;
 public:
     EFPDF(std::string arguments, std::string inputPath, std::string outputPath, std::string histogramPath);
     ~EFPDF();
